requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
et-xmlfile==1.1.0
googledrivedownloader==0.4
greenlet==1.1.2
idna==3.3
openpyxl==3.0.9
psycopg2==2.9.3
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
urllib3==1.26.9
